Pruebas unitarias V1

+++ ProyectoGestionV2/ProyectoGestion/src/main/java/com/unicauca/proyectogestion/domain/Usuario.java
@@ -3,13 +3,13 @@ package com.unicauca.proyectogestion.domain;
 public class Usuario {
     private String nombres;
     private String apellidos;    
-    private int celular;
+    private double celular;
     private EnumProgramas programa;
     private EnumRoles rol;
     private String email;    
     private String contrasenia;    
 
-    public Usuario(String nombres, String apellidos, int celular, EnumProgramas programa, EnumRoles rol, String email, String contrasenia) {
+    public Usuario(String nombres, String apellidos, double celular, EnumProgramas programa, EnumRoles rol, String email, String contrasenia) {
         this.nombres = nombres;
         this.apellidos = apellidos;
         this.celular = celular;
@@ -35,11 +35,11 @@ public class Usuario {
         this.apellidos = apellidos;
     }
 
-    public int getCelular() {
+    public double getCelular() {
         return celular;
     }
 
-    public void setCelular(int celular) {
+    public void setCelular(double celular) {
         this.celular = celular;
     }
